[#9] Chore: 프로젝트 디렉토리 구성, 불필요한 주석 제거
