--- todo-list-ios/TODOList/TODOList/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  TODOList
-//
-//  Created by 이준우 on 2022/04/04.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
